fix(pinia): rename store dir to stores
